Feat: check existed files for get stock data(kospi)

+++ stock_data/get_stock_data.py
@@ -9,15 +9,21 @@ print("Cybos 연결 준비")
 cpUtil = CpUtil()
 cpUtil.check_server_type()
 
-
-kospi_list = cpUtil.get_stock_list(CPE_MARKET_KIND.KOSPI)
-print("KOSPI 종목:")
-print(kospi_list[:5])
-
 cpSysDib = CpSysDib()
 
-# Kospi 5분봉 데이터 불러오기 & csv 파일 저장
-os.makedirs("./data/KOSPI",exist_ok=True)
-for code in tqdm(kospi_list[:3], desc="Get KOSPI", position=0):
+# KOSPI 5분봉 데이터 불러오기 & csv 파일 저장
+# 기존에 불러온 데이터 있는지 확인
+if os.path.exists("./data/KOSPI"):
+    existed_files = os.listdir("./data/KOSPI")
+    existed_files = {filename.split('.')[0] for filename in existed_files}
+else:
+    os.makedirs("./data/KOSPI",exist_ok=True)
+    existed_files = {}
+
+kospi_list = cpUtil.get_stock_list(CPE_MARKET_KIND.KOSPI)
+for code in tqdm(kospi_list, desc="Get KOSPI", position=0):
+    if code in existed_files:
+        continue
     tmp_df = cpSysDib.get_stock_data_p(code, '202006', '202506')
     tmp_df.to_csv(f"./data/KOSPI/{code}.csv", sep=',', index=False)
+    existed_files.add(code)
